audit-tumble-engine-bitexact-literals: s2 wrap 11 f32 literal sites, wire check into bun run check
Wrap all 11 sites the sweep flagged into rule 1a's fround-correct form
(f32(f32(lit) * operand), matching mesh.ts:646's sanctioned pattern).
Three constants move values to match the C reference:
  - core.ts:23 OVERLAP_SLOP: 0.0004999999655410647 -> 0.0005000000237487257
  - hull.ts:1410 minH: 0.0009999999310821295 -> 0.0010000000474974513
  - distance.ts:1154 kToleranceSquared: 0.0024999999441206455 -> 0.002500000176951289
The other 8 are correct by coincidence (value unchanged after wrapping).

Wire scripts/check-tumble-fp.ts into bun run check (package.json check chain).
Flip the arm's pinned count from 11 to 0; keep S1b's per-leg composition arms
and the trig allowlist (exactly two entries, keyed on file+deviation) intact.

Fixtures re-read: step.fixture.ts 55/0 both before and after — no fixture hash
shifted. The distance.ts:1154 differential: makeSeparationFunction is never
called during the fixture run (timeOfImpact exits early: Overlapped/Hit/Failed),
so the SeparationType.Vertices branch is not exercised — no flip, by the numbers
(solveContinuous called 163x, makeSeparationFunction called 0x).

+++ packages/shallot/src/standard/tumble/engine/geometry.ts
@@ -331,7 +331,7 @@ export function rayCastCapsule(shape: Capsule, input: RayCastInput): CastOutput 
     const d = vec3.sub(c2, c1);
 
     // Fall back to sphere if the capsule is short
-    const tol = f32(0.01 * LINEAR_SLOP);
+    const tol = f32(f32(0.01) * LINEAR_SLOP);
     const lengthSquared = vec3.lengthSq(d);
     if (lengthSquared < f32(tol * tol)) {
         const sphereCenter = vec3.scale(0.5, vec3.add(shape.center1, shape.center2));
